+++ item.go
@@ -30,3 +30,33 @@ func (s *Service) GetItems(GroupUUID uuid.UUID) (Items, error) {
 
 	return resp.Data.Items, nil
 }
+
+// CreateItem creates a new item in a group based on the ItemPayload.
+func (s *Service) CreateItem(item ItemPayload) (Item, error) {
+	s.URL.Path = "/item/"
+	p, err := marshalReader(item)
+	if err != nil {
+		return Item{}, err
+	}
+
+	res, e := s.DoRequest("POST", s.URL, nil, nil, p)
+	if e != nil {
+		return Item{}, e
+	}
+
+	type data struct {
+		Item Item `json:"item"`
+	}
+
+	resp := struct {
+		Message string `json:"message"`
+		Data    data   `json:"data"`
+	}{}
+
+	err = marshalResponse(201, res, &resp)
+	if err != nil {
+		return Item{}, err
+	}
+
+	return resp.Data.Item, nil
+}