+++ item_test.go
@@ -73,3 +73,91 @@ func TestService_GetItems(t *testing.T) {
 		})
 	}
 }
+
+func TestService_CreateItem(t *testing.T) {
+	tests := []struct {
+		name     string
+		payload  ItemPayload
+		exchange *microtest.Exchange
+		uri      string
+		item     Item
+		e        error
+	}{
+		{
+			name: "403 permission required",
+			payload: ItemPayload{
+				GroupUUID: uuid.MustParse("40355dba-0923-43a6-83d5-c9b6680edd2e"),
+				Name:      "test item",
+			},
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 403,
+					Body:   noPermission,
+				},
+			},
+			uri:  "/item/",
+			item: Item{},
+			e:    errors.New("no permission"),
+		},
+		{
+			name: "404 group not found",
+			payload: ItemPayload{
+				GroupUUID: uuid.MustParse("f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8"),
+				Name:      "test item",
+			},
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 404,
+					Body:   notFound,
+				},
+			},
+			uri:  "/item/",
+			item: Item{},
+			e:    errors.New("not found"),
+		},
+		{
+			name: "201 successful",
+			payload: ItemPayload{
+				GroupUUID: uuid.MustParse("f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8"),
+				Name:      "test item",
+			},
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 201,
+					Body:   string(responseItemNew),
+				},
+			},
+			uri:  "/item/",
+			item: testItemNew,
+			e:    nil,
+		},
+	}
+
+	s := NewService(Config{})
+	ms := microtest.MockServer(s)
+	defer ms.Server.Close()
+
+	for i, tc := range tests {
+		name := fmt.Sprintf("%d %s", i, tc.name)
+		t.Run(name, func(t *testing.T) {
+			ms.Append(tc.exchange)
+
+			item, e := s.CreateItem(tc.payload)
+
+			// ensure errors match as expected
+			if NotEqualError(e, tc.e) {
+				t.Errorf("expected error %v, got %v", tc.e, e)
+			}
+
+			// ensure items match as expected
+			if EqualItem(item, tc.item) == false {
+				t.Errorf("expected item\n%+v\ngot\n%+v", tc.item, item)
+			}
+
+			// ensure the correct uri was called
+			if tc.uri != tc.exchange.Request.RequestURI {
+				t.Errorf("expected uri %s, got %s", tc.uri, tc.exchange.Request.RequestURI)
+			}
+		})
+	}
+}